Major Changes
Fixed css issues with sign-in and sign-up so they are centered now regardless of resizing. Restyled the home page for mobile, restyled add task section. Added ability to see task information, added ability to delete a task once it has reached the done section. Styled the profile page to show information in center of page with a basic profile image that is not connected to the account, just for show.